Move json to core
