+++ src/json.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2015 The AMP HTML Authors. All Rights Reserved.
+ * Copyright 2021 The AMP HTML Authors. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,132 +14,8 @@
  * limitations under the License.
  */
 
-/**
- * @fileoverview This module declares JSON types as defined in the
- * {@link http://json.org/}.
- */
-
 import {childElementsByTag, isJsonScriptTag} from './dom';
-import {isObject} from './core/types';
-
-// NOTE Type are changed to {*} because of
-// https://github.com/google/closure-compiler/issues/1999
-
-/**
- * JSON scalar. It's either string, number or boolean.
- * @typedef {*} should be string|number|boolean|null
- */
-let JSONScalarDef;
-
-/**
- * JSON object. It's a map with string keys and JSON values.
- * @typedef {*} should be !Object<string, ?JSONValueDef>
- */
-let JSONObjectDef;
-
-/**
- * JSON array. It's an array with JSON values.
- * @typedef {*} should be !Array<?JSONValueDef>
- */
-let JSONArrayDef;
-
-/**
- * JSON value. It's either a scalar, an object or an array.
- * @typedef {*} should be !JSONScalarDef|!JSONObjectDef|!JSONArrayDef
- */
-let JSONValueDef;
-
-/**
- * @typedef {{
- *   YOU_MUST_USE: string,
- *   jsonLiteral: function(),
- *   TO_MAKE_THIS_TYPE: string,
- * }}
- */
-let InternalJsonLiteralTypeDef;
-
-/**
- * Recreates objects with prototype-less copies.
- * @param {!JsonObject} obj
- * @return {!JsonObject}
- */
-export function recreateNonProtoObject(obj) {
-  const copy = Object.create(null);
-  for (const k in obj) {
-    if (!hasOwnProperty(obj, k)) {
-      continue;
-    }
-    const v = obj[k];
-    copy[k] = isObject(v) ? recreateNonProtoObject(v) : v;
-  }
-  return /** @type {!JsonObject} */ (copy);
-}
-
-/**
- * Returns a value from an object for a field-based expression. The expression
- * is a simple nested dot-notation of fields, such as `field1.field2`. If any
- * field in a chain does not exist or is not an object or array, the returned
- * value will be `undefined`.
- *
- * @param {!JsonObject} obj
- * @param {string} expr
- * @return {*}
- */
-export function getValueForExpr(obj, expr) {
-  // The `.` indicates "the object itself".
-  if (expr == '.') {
-    return obj;
-  }
-  // Otherwise, navigate via properties.
-  const parts = expr.split('.');
-  let value = obj;
-  for (let i = 0; i < parts.length; i++) {
-    const part = parts[i];
-    if (
-      part &&
-      value &&
-      value[part] !== undefined &&
-      hasOwnProperty(value, part)
-    ) {
-      value = value[part];
-      continue;
-    }
-    value = undefined;
-    break;
-  }
-  return value;
-}
-
-/**
- * Simple wrapper around JSON.parse that casts the return value
- * to JsonObject.
- * Create a new wrapper if an array return value is desired.
- * @param {*} json JSON string to parse
- * @return {?JsonObject} May be extend to parse arrays.
- */
-export function parseJson(json) {
-  return /** @type {?JsonObject} */ (JSON.parse(/** @type {string} */ (json)));
-}
-
-/**
- * Parses the given `json` string without throwing an exception if not valid.
- * Returns `undefined` if parsing fails.
- * Returns the `Object` corresponding to the JSON string when parsing succeeds.
- * @param {*} json JSON string to parse
- * @param {function(!Error)=} opt_onFailed Optional function that will be called
- *     with the error if parsing fails.
- * @return {?JsonObject} May be extend to parse arrays.
- */
-export function tryParseJson(json, opt_onFailed) {
-  try {
-    return parseJson(json);
-  } catch (e) {
-    if (opt_onFailed) {
-      opt_onFailed(e);
-    }
-    return null;
-  }
-}
+import {parseJson} from './core/types/object/json';
 
 /**
  * Helper method to get the json config from an element <script> tag
@@ -163,109 +39,4 @@ export function getChildJsonConfig(element) {
   } catch (unusedError) {
     throw new Error('Failed to parse <script> contents. Is it valid JSON?');
   }
-}
-
-/**
- * Deeply checks strict equality of items in nested arrays and objects.
- *
- * @param {JSONValueDef} a
- * @param {JSONValueDef} b
- * @param {number} depth The maximum depth. Must be finite.
- * @return {boolean}
- * @throws {Error} If depth argument is not finite.
- */
-export function deepEquals(a, b, depth = 5) {
-  if (!isFinite(depth) || depth < 0) {
-    throw new Error('Invalid depth: ' + depth);
-  }
-  if (a === b) {
-    return true;
-  }
-  /** @type {!Array<{a: JSONValueDef, b: JSONValueDef, depth: number}>} */
-  const queue = [{a, b, depth}];
-  while (queue.length > 0) {
-    const {a, b, depth} = queue.shift();
-    // Only check deep equality if depth > 0.
-    if (depth > 0) {
-      if (typeof a !== typeof b) {
-        return false;
-      } else if (Array.isArray(a) && Array.isArray(b)) {
-        if (a.length !== b.length) {
-          return false;
-        }
-        for (let i = 0; i < a.length; i++) {
-          queue.push({a: a[i], b: b[i], depth: depth - 1});
-        }
-        continue;
-      } else if (a && b && typeof a === 'object' && typeof b === 'object') {
-        const keysA = Object.keys(/** @type {!Object} */ (a));
-        const keysB = Object.keys(/** @type {!Object} */ (b));
-        if (keysA.length !== keysB.length) {
-          return false;
-        }
-        for (let i = 0; i < keysA.length; i++) {
-          const k = keysA[i];
-          queue.push({a: a[k], b: b[k], depth: depth - 1});
-        }
-        continue;
-      }
-    }
-    // If we get here, then depth == 0 or (a, b) are primitives.
-    if (a !== b) {
-      return false;
-    }
-  }
-  return true;
-}
-
-/**
- * @param {*} obj
- * @param {string} key
- * @return {boolean}
- */
-function hasOwnProperty(obj, key) {
-  if (obj == null || typeof obj != 'object') {
-    return false;
-  }
-  return Object.prototype.hasOwnProperty.call(
-    /** @type {!Object} */ (obj),
-    key
-  );
-}
-
-/**
- * This helper function handles configurations specified in a JSON format.
- *
- * It allows the configuration is to be written in plain JS (which has better
- * dev ergonomics like comments and trailing commas), and allows the
- * configuration to be transformed into an efficient JSON-parsed representation
- * in the dist build. See https://v8.dev/blog/cost-of-javascript-2019#json
- *
- * @param {!Object} obj
- * @return {!JsonObject}
- */
-export function jsonConfiguration(obj) {
-  return /** @type {!JsonObject} */ (obj);
-}
-
-/**
- * This converts an Object into a suitable type to be used in `includeJsonLiteral`.
- * This doesn't actually do any conversion, it only changes the closure type.
- *
- * @param {!Object|!Array|string|number|boolean|null} value
- * @return {!InternalJsonLiteralTypeDef}
- */
-export function jsonLiteral(value) {
-  return /** @type {!InternalJsonLiteralTypeDef} */ (value);
-}
-
-/**
- * Allows inclusion of a variable (that's wrapped in a jsonLiteral
- * call) to be included inside a jsonConfiguration.
- *
- * @param {!InternalJsonLiteralTypeDef} value
- * @return {*}
- */
-export function includeJsonLiteral(value) {
-  return value;
 }
